V0.0.6 将存款类型float改为double类型 新增存入存款查看存款功能 优化登录账号功能

+++ db.py
@@ -15,7 +15,7 @@ def init_db():
         (
             id INT PRIMARY KEY AUTO_INCREMENT,
             bank_code TEXT NOT NULL,    
-            deposit FLOAT NOT NULL,
+            deposit DOUBLE NOT NULL,
             user_name TEXT NOT NULL,
             password INT NOT NULL, 
             created_at BIGINT NOT NULL,
